Move warning into diagnostics directory
