+++ src/diagnostics/warning.js
